@@ -1,5 +1,5 @@
-const ArgvParser = require("./ArgvParser");
-const ComparisonTestsRunner = require("./ComparisonTestsRunner");
+import { ArgvParser } from "./ArgvParser";
+import { ComparisonTestsRunner } from "./ComparisonTestsRunner";
 
 const testNames = new ArgvParser(process.argv).parseCommandNames();
 const integrationTests = new ComparisonTestsRunner("test/integration", testNames);
